+++ database/migrations/2022_02_03_000752_create_histories_table.php
@@ -18,12 +18,11 @@ class CreateHistoriesTable extends Migration
             $table->string('des_ar');
             $table->string('des_en');
             $table->integer('member_id');
-            $table->integer('store_id');
+            $table->unsignedBigInteger('store_id');
             $table->integer('status')->default(0);
             $table->timestamps();
 
             $table->foreign('store_id')->references('id')->on('stores')->onDelete('cascade')->onUpdate('cascade');
-            // $table->foreign('member_id')->references('id')->on('users')->onDelete('SET NULL')->onUpdate('SET NULL');
         });
     }
